requirements.txt
streamlit
easyocr
langdetect
deep-translator
Pillow
